# backend/src/main/java/com/waimaicps/common/ApiResponse.java
package com.waimaicps.common;

public record ApiResponse<T>(String code, String message, T data, String requestId) {
    public static <T> ApiResponse<T> success(T data, String requestId) {
        return new ApiResponse<>("SUCCESS", "ok", data, requestId);
    }
}