# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Waimai CPS API
  version: 1.0.0
servers:
  - url: https://api.example.com/api/v1
paths:
...
  /admin/auth/login:
    post: {summary: 租户管理员登录, security: [], responses: {"200": {$ref: "#/components/responses/Success"}, "401": {$ref: "#/components/responses/Error"}}}
  /admin/auth/csrf:
    get: {summary: 初始化 CSRF Cookie, security: [{adminCookie: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /admin/auth/logout:
    post: {summary: 租户管理员退出, security: [{adminCookie: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /admin/dashboard:
    get: {summary: 租户运营概览, security: [{adminCookie: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /platform/tenants:
    get: {summary: 平台管理员查询租户, security: [{adminCookie: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
components:
  securitySchemes:
    miniBearer: {type: http, scheme: bearer, bearerFormat: opaque-random-token}
    adminCookie: {type: apiKey, in: cookie, name: ADMIN_SESSION}
...
  schemas:
...
    ApiResponse: {type: object, required: [code, message, requestId], properties: {code: {type: string}, message: {type: string}, data: {}, requestId: {type: string, format: uuid}}}
  responses:
    Success: {description: 成功, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}
    Error: {description: 业务错误, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}